Full implementation of the report on SQL
This commit adds a rough implementation of a CF usage report
implemented in SQL directly.

The goal is to generate the reports as done in
the repository usage-reports in the most quick and simple way possible:

 * We rely on a series of SQL queries to sum all the events from a copy
   of the data from the table `app_usage_events` of the CF
   CloudController, which is the data exposed by the endpoint
   `/app_usage_events`[1]

 * We dump all the data from the environments using bosh2+ssh+psql and
   import it in a PostgresSQL running on docker.

 * We run the queries on the postgresql running on a local docker
   container, and write the result as CSV in the file.

See the commited README.md, the image and the implementation for
details of the implementation.

All the process to download the data and generate the report
is driven from a Makefile

[1] https://docs.cloudfoundry.org/running/managing-cf/usage-events.html).