Aggregate applications running but with no events
If an application has been running for all the month without being
restarted, scaled, deployed or stopped, it will have no events
in the usage events.

But these applications have been running for all the month.

Because that, we must also take them into account. To do so
we query information about the existing applications, instances,
memory and state, when fetching all the data in one table
called `existing_apps`

When generating the report, we query this table to check if there
is any running app without events. If there is any, we
sum its usage for all the month.