Carried usage is computed as fake events
We want to simplify the code and keep the aggregation logic only
in one query. This will help when implementing the multi-tier
logic.

This might also help to understand how the logic works.

For that, in order to compute the carried usage from previous month,
that is:
 - Usage before the first event
 - Usage for apps that have no events

We will create some views with "fake events" representing the
application starting at the beginning of the month in each case.

That way, we keep the aggregation logic only in
the query per_app_usage_in_secs.