Refactor the queries to split events in metrics
Each app_usage_event actually refers to two different "metrics"

 - The application have started to consume X mem * Y instances
 - The application have stopped to consume previous X mem * Y instances.

If we generate 2 metrics for each event, it simplifies the rest of the
queries being computed.

The "previous" metric has to set instances to a negative value so that
is later substracted from the total.

In this model, the carried usage would be "fake metrics" added
at the beginning of the window.