Implement memory usage tiers in report
We want to generate a report that would break down the total
consumption in size of the instances in memory.

We define 4x tiers:
 - 0GB-0.5GB
 - 0.5GB-1GB
 - 1GB-2GB
 - 2GB+

In order to compute the tiers, we create a new materialized view,
app_usage_metrics_tiers, which will simply add a tier to classify
each usage. That table will gather the metrics from the original
app_usage and the carried.

In the next queries, we group by org+space+app+tier. This way
we get the values aggregated by tier.

Finally we add another view that would do a outer join of every single
org+space with each one of the tiers. This allows us to create one
row with the usage for each org+space with total,tier1,tier2...