Add a table to pin the report window
We want to easily be able to declare the time window to be
used in the report.

But in postgres there is no easy way to use global variables or
parameters.

The best solution is to create a new table with only one row and
the 3 values:
 - t_start
 - t_end
 - t_interval in secs

To compute it, we set the base date for t_end in a subquery.

"interval '1 month'" has a contextual behaviour. It can be resolved
to 28, 29, 30 or 31 depending which date is substracted. If it is
not substracted, it will be based on now().

But we want it to be based on the value of t_end and t_start. Because
that we compute it as:

 base_date - (base_date - interval '1 month') '1 month' AS t_interval