Be smart with usage carried outside of the window
When computing the usage, we must take into account all
the applications that were running during the window, but
that do not have events within the window (e.g. were not started,
stopped or scaled)

We have to cover the following cases:

 - No events within the window, but events in all the events. 2 cases:
   - If there is a event AFTER => usage as previous for all window
   - If there is a event BEFORE (but not after)  => usage as current

 - No events at all, but app is running:
   - Add usage if updated_at && create_at < window.t_start

We add a compressive test for these cases.

This commit fixes the bug that we might be overcharging if
an app was started after the timewindow.

Signed-off-by: Ryan Grenz <ryan.grenz@fil.com>