Test CSV header in specific test, ignore in other
We add a test to specifically test the header of the CSV report.

Then we strip off the header when running the rest of the tests.

This way change the header of the CSV would be easy, but it is
still being tested, to ensure that it is not changed accidentally.