Remove logic to download all usage data
We want to release to the open this repository, and the
specifics of how we download all the data from our envs is
specific to our installation.

We remove the download_all_usage.sh script and move it to
a script in the utils pipeline definition.

We keep a target in the makefile: download_all_usage which would call
the script in our utils pipeline, but that script is not commited here.